+++ resources/js/views/Shared.vue
@@ -31,7 +31,7 @@
             <MobileToolbar />
 
             <!--File list & info sidebar-->
-            <div class="flex space-x-3 lg:overflow-hidden grow">
+            <div class="flex space-x-3 lg:overflow-hidden grow" @drop.stop.prevent="uploadDroppedItems($event)" @dragenter.prevent @dragover.prevent>
                 <router-view id="file-view" class="relative w-full" :key="$route.fullPath" />
 
                 <InfoSidebar v-if="isVisibleSidebar" />
@@ -81,7 +81,7 @@ export default {
         Alert,
     },
     computed: {
-        ...mapGetters(['isVisibleSidebar', 'sharedDetail', 'config']),
+        ...mapGetters(['isVisibleSidebar', 'sharedDetail', 'config', 'currentFolder']),
     },
     data() {
         return {
@@ -89,6 +89,9 @@ export default {
         }
     },
     methods: {
+		uploadDroppedItems(event) {
+			this.$uploadDraggedFiles(event, this.currentFolder?.data.id)
+		},
         contextMenu(event, item) {
             events.$emit('context-menu:show', event, item)
         },