+++ resources/views/index.blade.php
@@ -1,6 +1,31 @@
 @php
 
     try {
+        // Bcmath Extension
+        $storageDefaultSpaceFormatted = isset($settings->default_max_storage_amount)
+            ? format_gigabytes($settings->default_max_storage_amount)
+            : format_gigabytes(5);
+
+        $uploadLimit = isset($settings->upload_limit)
+            ? format_bytes($settings->upload_limit)
+            : 'undefined';
+
+        $chunkSize = isset($settings->chunk_size)
+            ? format_bytes($settings->chunk_size)
+            : format_bytes(64);
+
+        $uploadLimitFormatted = isset($settings->upload_limit)
+            ? format_megabytes($settings->upload_limit)
+            : null;
+
+    } catch (MissingExtensionException $exception) {
+        $storageDefaultSpaceFormatted = '5GB';
+        $uploadLimit = 'undefined';
+        $uploadLimitFormatted = 5;
+        $chunkSize = 64000000;
+    }
+
+    try {
         // User
         $isUser = auth()->check();
         $user = Auth::user();
@@ -77,11 +102,11 @@
 
 			storageLimit: {{ $settings->storage_limitation ?? 1 }},
 			storageDefaultSpace: {{ $settings->default_max_storage_amount ?? 5 }},
-			storageDefaultSpaceFormatted: '{{ isset($settings->default_max_storage_amount) ? format_gigabytes($settings->default_max_storage_amount) : format_gigabytes(5) }}',
+			storageDefaultSpaceFormatted: '{{ $storageDefaultSpaceFormatted }}',
 			mimetypesBlacklist: '{{ isset($settings->mimetypes_blacklist) ? $settings->mimetypes_blacklist: null}}',
-			uploadLimit: {{ isset($settings->upload_limit) ? format_bytes($settings->upload_limit) : 'undefined' }},
-			uploadLimitFormatted: '{{ isset($settings->upload_limit) ? format_megabytes($settings->upload_limit) : null }}',
-			chunkSize: {{ isset($settings->chunk_size) ? format_bytes($settings->chunk_size) : format_bytes(64) }},
+			uploadLimit: {{ $uploadLimit }},
+			uploadLimitFormatted: '{{ $uploadLimitFormatted }}',
+			chunkSize: {{ $chunkSize }},
 
 			isAuthenticated: {{ $isUser ? 1 : 0 }},