+++ src/Domain/Localization/Actions/UpdateLanguageTranslationsAction.php
@@ -2,6 +2,7 @@
 namespace Domain\Localization\Actions;
 
 use DB;
+use Artisan;
 
 class UpdateLanguageTranslationsAction
 {
@@ -10,9 +11,11 @@ class UpdateLanguageTranslationsAction
         collect($list)
             ->each(
                 fn (...$item) => DB::table('language_translations')
-            ->where('lang', 'en')
-            ->where('key', $item[1])
-            ->update(['value' => $item[0]])
+                    ->where('lang', 'en')
+                    ->where('key', $item[1])
+                    ->update(['value' => $item[0]])
             );
+
+        Artisan::call('cache:clear');
     }
 }