+++ changelog.md
@@ -2,8 +2,10 @@
 #### Release date: 9. May 2022
 - Ability to manually synchronize plans in fixed subscription type
 - Improved sanitization for .env values to prevent crash your app
-- Improved reCaptcha validation error logs
+- Improved reCaptcha validation errors
 - Fixed issue when upload doesn't start after you drag the file into empty view
+- Fixed issue when homepage flash to sign in screen after the homepage was disabled in admin panel
+- Fixed trash navigator issue
 
 ## Version 2.1.2
 #### Release date: 8. May 2022